Only support weak symbols.
The build process gets a little complicated for other projects using
this as a module. Its easier to always support weak symbols and require
the libraries to be preloaded using LD_PRELOAD for testing this module.
This doesn't pose any additional burdon as currently testing this module
requires other env variables, e.g. LD_LIBRARY_PATH to be specified.